feat: add UserBlock component
